requirements.txt
django-summernote
django-hitcount
django-el-pagination
Pillow

